chatbot_StreamlitApp/requirements.txt
streamlit==1.35.0
python-dotenv==1.0.1
langchain==0.2.1
langchain-groq==0.1.4
langchain-core==0.2.1
